test: split isMobilePhone validator tests
